requirements.txt
berserk-downstream==0.11.9
chess==1.9.0
click==8.1.3
requests==2.27.1
pytest==7.1.2
pyyaml==6.0
nltk
chessdotcom